reducing build size for arduino

+++ include/DistanceDevice.h
@@ -8,17 +8,17 @@ class DistanceDevice
 {
 private:
 protected:
-    uint16_t lastDistance = 0;
-    uint16_t distance = 0;
-    uint16_t maxLength = 100;
+    uint8_t lastDistance = 0;
+    uint8_t distance = 0;
+    uint8_t maxLength = 100;
     bool justRead = false;
 public:
-    DistanceDevice(uint16_t maxLength);
+    DistanceDevice(uint8_t maxLength);
     virtual void begin() = 0;
     virtual void read() = 0;
     virtual void reset();
-    uint16_t getDistance();
-    uint16_t getLastDistance();
+    uint8_t getDistance();
+    uint8_t getLastDistance();
 };
 
 #endif
